+++ workspace/index.js
@@ -1,9 +1,17 @@
-var module = angular.module('impacWorkspace', ['maestrano.impac']);
+var module = angular.module('impacWorkspace', ['maestrano.impac', 'toastr']);
 
 // Configuration impac-angular lib on module impacWorkSpace run.
-module.run(function ($log, $window, $q, $http, ImpacLinking, ImpacRoutes, ImpacTheming) {
+module.run(function ($log, $window, $q, $http, ImpacLinking, ImpacRoutes, ImpacTheming, ImpacDeveloper, toastr) {
 
-  // TODO: set-up server to enable local $http calls to setting.json
+  //--------------------------------------------------------
+  // Start editing
+  //--------------------------------------------------------
+  // PLEASE DO NOT COMMIT CHANGES TO THIS FILE.
+  // You can use `git add -p ./` to selectively add files to your git index.
+  // -------------------------------------------------------
+
+  // Credentials and endpoints
+  //------------------------------------------------
   var settings = {
     mno_url: 'https://uat.maestrano.io',
     impac_url: 'https://api-impac-uat.maestrano.io',
@@ -11,9 +19,37 @@ module.run(function ($log, $window, $q, $http, ImpacLinking, ImpacRoutes, ImpacT
     api_secret: ''
   };
 
+  // Stub widget templates - add new widgets!
+  //------------------------------------------------
+  // Widget templates are stored on Maestrano API, stub your
+  // new widget templates here, and contact us for permenant additions or
+  // changes to existing widget templates.
+  var widgetsTemplates = [
+    // Example widget template, please see documentation "How-To: Create a Widget" for available
+    // options, and importantly, note how the path & metadata.template attributes work.
+    // --
+    // {
+    //   path: 'accounts/assets_summary',
+    //   name: 'Assets / Liabilities summary',
+    //   metadata: { template: 'accounts/assets_liability_summary' },
+    //   desc: 'Compare Assets and Liabilities accounts',
+    //   icon: 'pie-chart',
+    //   width: 3
+    // },
+  ];
+  // --
+
+
+  //--------------------------------------------------------
+  // Do not edit below unless you know what you're doing.
+  //--------------------------------------------------------
+  // Check credentials have been provided
   if (!settings.api_key || !settings.api_secret) {
-    fail('missing authentication credentials!');
+    fail('Missing authentication credentials!');
   }
+
+  // Impac-angular configurations.
+  // -------------------------------------------------------
 
   // encodes a base64 string - Basic Authentication.
   var credentials = $window.btoa(settings.api_key + ':' + settings.api_secret);
@@ -38,6 +74,12 @@ module.run(function ($log, $window, $q, $http, ImpacLinking, ImpacRoutes, ImpacT
     }
   });
 
+  // Configure the ImpacDeveloper service options.
+  ImpacDeveloper.configure({
+    status: true,
+    widgetsTemplates: widgetsTemplates || []
+  });
+
   // Link core callbacks required for impac-angular lib to run.
   ImpacLinking.linkData({
     organizations: function () {
@@ -53,14 +95,16 @@ module.run(function ($log, $window, $q, $http, ImpacLinking, ImpacRoutes, ImpacT
       .then(function (response) {
         var organizations = (response.data || []);
         return { organizations: organizations, currentOrgId: (organizations[0].id || null) };
-      },
-        fail
-      );
+      }, function () {
+        var msg = 'Unable to retrieve Organizations';
+        fail(msg);
+        return $q.reject(msg);
+      });
   }
 
-  function fail(err) {
-    $log.error('workspace/index.js ERROR: ', err);
-    return {};
+  function fail(msg) {
+    $log.error('workspace/index.js Error: ' + msg);
+    toastr.error(msg, 'Error');
   }
 
 });